fix(coding-agent): keep auto-title prompts private
